Refactor polishing

+++ src/Application/Handler/DispatchArea/DeleteDispatchAreaHandler.php
@@ -3,43 +3,44 @@
 namespace App\Application\Handler\DispatchArea;
 
 use App\Application\Contract\HandlerInterface;
+use App\Application\Exception\DispatchAreaNotEmptyException;
+use App\Application\Traits\EventDispatcherTrait;
 use App\Domain\Command\DispatchArea\DeleteDispatchAreaCommand;
+use App\Domain\Entity\DispatchArea;
 use App\Domain\Event\DispatchArea\DispatchAreaDeletedEvent;
 use App\Repository\DispatchAreaRepository;
 use App\Repository\StateRepository;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 class DeleteDispatchAreaHandler implements HandlerInterface
 {
+    use EventDispatcherTrait;
+
     private DispatchAreaRepository $dispatchAreaRepository;
 
     private StateRepository $stateRepository;
 
-    private EventDispatcherInterface $dispatcher;
-
-    public function __construct(DispatchAreaRepository $dispatchAreaRepository, StateRepository $stateRepository, EventDispatcherInterface $dispatcher)
+    public function __construct(DispatchAreaRepository $dispatchAreaRepository, StateRepository $stateRepository)
     {
         $this->dispatchAreaRepository = $dispatchAreaRepository;
         $this->stateRepository = $stateRepository;
-
-        $this->dispatcher = $dispatcher;
     }
 
     public function __invoke(DeleteDispatchAreaCommand $command): void
     {
+        /** @var DispatchArea $area */
         $area = $this->dispatchAreaRepository->getById($command->getId());
 
-        // @TODO Check if DispatchArea includes any hospitals!
+        if (!$area->getHospitals()->isEmpty()) {
+            throw new DispatchAreaNotEmptyException(sprintf('Cannot delete Area "%s" (#%d) as it still contains Hospitals', $area->getName(), $area->getId()));
+        }
 
         $this->dispatchAreaRepository->delete($area);
 
-        $state = $this->stateRepository->getById($area->getState()->getId());
+        $state = $this->stateRepository->findOneBy(['id' => $area->getState()->getId()]);
         $state->removeDispatchArea($area);
 
         $this->stateRepository->save();
 
-        $event = new DispatchAreaDeletedEvent($area);
-
-        $this->dispatcher->dispatch($event, DispatchAreaDeletedEvent::NAME);
+        $this->dispatchEvent(new DispatchAreaDeletedEvent($area));
     }
 }
